# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ficha Digital API",
        "description": "API para el sistema de gestión de solicitudes y fichas digitales",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/provincias": {
            "get": {
                "tags": [
                    "Datos de Referencia"
                ],
                "summary": "Listar todas las provincias",
                "operationId": "126248eadc980bc680feb7d63e5215fc",
                "responses": {
                    "200": {
                        "description": "Lista de provincias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Provincias"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "codigo": {
                                                        "type": "string",
                                                        "example": "BA"
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Buenos Aires"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/localidades": {
            "get": {
                "tags": [
                    "Datos de Referencia"
                ],
                "summary": "Listar localidades con filtros",
                "description": "Lista localidades filtradas por provincia y/o búsqueda (mínimo 3 caracteres)",
                "operationId": "d988e7b51d91a4a8fd84ec141ab59cea",
                "parameters": [
                    {
                        "name": "provincia_id",
                        "in": "query",
                        "description": "ID de la provincia para filtrar",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "busqueda",
                        "in": "query",
                        "description": "Texto de búsqueda (mínimo 3 caracteres)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Mar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de localidades",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Localidades"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "provincia_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Mar del Plata"
                                                    },
                                                    "provincia": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "codigo": {
                                                                "type": "string",
                                                                "example": "BA"
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Buenos Aires"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "La búsqueda debe tener al menos 3 caracteres"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/paises": {
            "get": {
                "tags": [
                    "Datos de Referencia"
                ],
                "summary": "Listar todos los países",
                "operationId": "190552ec58c8ece1ffb0433fb5408c9e",
                "responses": {
                    "200": {
                        "description": "Lista de países",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Países"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Argentina"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}